move: rename MainPageContainer -> PageContainer, move MainPageStyle -> GlobalStyle
