+++ src/App.jsx
@@ -1,4 +1,4 @@
-import GlobalStyle from "components/styles/GlobalStyle";
+import { GlobalStyle } from "components/styles/GlobalStyle";
 import Router from "shared/Router";
 
 function App() {
