--- src/components/styles/MainPageStyle.jsx
@@ -1,17 +1,0 @@
-import styled from "styled-components";
-
-const MainPageContainer = styled.div`
-  display: flex;
-  justify-content: space-evenly;
-`;
-
-const SectionContainer = styled.div`
-  margin: 20px;
-`;
-
-const SectionTitle = styled.h2`
-  font-size: 20px;
-  font-weight: 900;
-`;
-
-export { MainPageContainer, SectionContainer, SectionTitle };
